[FLINK-11466][e2e] Replace the FlinkDistribution with LocalStandaloneFlinkResource
