+++ flink-end-to-end-tests/flink-end-to-end-tests-common/src/main/java/org/apache/flink/tests/util/End2EndUtil.java
@@ -30,7 +30,7 @@ public class End2EndUtil {
 	private static final String END_TO_END_TESTS_MODULE = "/flink-end-to-end-tests/";
 
 	/**
-	 * Get the path of <flink-root-dir>/flink-end-to-end-tests based on the class loading.
+	 * Get the path of FLINK_ROOT_DIR/flink-end-to-end-tests based on the class loading.
 	 */
 	public static Path getEnd2EndModuleDir() {
 		URL url = End2EndUtil.class.getResource("End2EndUtil.class");
@@ -42,7 +42,7 @@ public class End2EndUtil {
 	}
 
 	/**
-	 * Generate a test directory path under the <flink-root-dir>/flink-end-to-end-tests/target, each time with a
+	 * Generate a test directory path under the FLINK_ROOT_DIR/flink-end-to-end-tests/target, each time with a
 	 * different directory.
 	 *
 	 * @return the test directory to put the temporary end-to-end test data.
